requirements.txt
pytorch >= 1.0.0
numpy >= 1.16.2
pandas >= 0.24.1
pyrubberband >= 0.3.0
librosa >= 0.6.3
pyyaml >= 3.13
mir_eval >= 0.5
pretty_midi >= 0.2.8
